+++ app/src/main/AndroidManifest.xml
@@ -31,9 +31,6 @@
             android:theme="@style/Theme.Bitsy.NoActionBar"
             android:screenOrientation="portrait"
             android:windowSoftInputMode="adjustPan"/>
-        <activity
-            android:name=".activities.SettingsActivity"
-            android:label="@string/title_settings"/>
     </application>
 
 </manifest>
