+++ app/src/main/res/layout/fragment_settings.xml
@@ -9,7 +9,7 @@
     android:paddingEnd="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="32dp"
-    tools:context=".activities.SettingsActivity">
+    tools:context=".fragments.SettingsFragment">
 
     <TextView
         android:layout_width="match_parent"
