+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -39,6 +39,12 @@
         tools:layout="@layout/fragment_transactions"/>
 
     <fragment
+        android:id="@+id/settings_dest"
+        android:name="cy.agorise.bitsybitshareswallet.fragments.SettingsFragment"
+        android:label="@string/title_settings"
+        tools:layout="@layout/fragment_settings"/>
+
+    <fragment
         android:id="@+id/receive_dest"
         android:name="cy.agorise.bitsybitshareswallet.fragments.ReceiveTransactionFragment"
         android:label="@string/title_receive"
